Move all navigation bar tweaking to viewWillAppear
This allows the system to properly animate the navigtion bar when it
pushes/pops view controllers. I’m not sure if all of these view
controllers are being used, but this is everything I could find that was
calling `-[UINavigationController setNavigationBarHidden:]`.

I took a method call out of a shared method and put it back into the
respective view controllers’ `-viewWillAppear:` method. If you want to put
it back in the shared method, you’ll need to capture the `animated`
parameter of the original call to `-viewWillAppear:`.